// apps/backend/src/dto/client.dto.ts
export type ClientDTO = {
  id: number;
  name: string;
  gender?: string;
  birthDate?: string;
  cpf?: string;
  rg?: string;
  email?: string;
  phone?: string;
  cep?: string;
  address?: string;
  number?: string;
  neighborhood?: string;
  city?: string;
  state?: string;
  complement?: string;
  created_at: string;
  company_id: string;
  active?: boolean;
  observation?: string;
};